Add workaround to support criticality
CredentialManager has an issue where it's incorrectly using the
requested elements instead of the candidate (critical) elements in the
matching phase. To work around this, we'll swap those in the Bundle send
to CredentialManager. Later, we reverse them once again so that the
provider gets the correct set of request elements.

The example app has been updated to send `family_name` as the only
critical element.

The `GetCredentialOption` constructor has been made `internal`, replaced
with a companion function `create()` which is needed to take care of the
swapping described above.