Fix support for critical elements
CredentialManager is incorrectly using the requested elements instead of
the candidate elements for querying the registry. This makes it
impossible to use the candidate and request bundles as designed.
Instead, we now set the requested elements in a non-standard Bundle key
and set the critical elements as both the candidate and request
elements. The public API is unchanged.

This also adds criticality to the example app, requesting only
`family_name` in the critical elements.